requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.2.5
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
distlib==0.3.1
Django==3.0.7
django-crispy-forms==1.9.1
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.7.0
isort==4.3.20
jdcal==1.4.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==7.1.2
psycopg2==2.8.3
psycopg2-binary==2.8.5
pylint==2.3.1
pyodbc==4.0.30
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.1
style==1.1.0
typed-ast==1.3.5
update==0.0.1
urllib3==1.25.6
virtualenv==20.0.30
wrapt==1.11.1
zipp==3.1.0